requirements.txt
# Pillar 3: Multi-Agent Creative Team
# Python 3.11+ required

# Core LangChain and LangGraph - Updated compatible versions
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-core>=0.3.17,<0.4.0

# OpenAI
openai==1.54.3

# Environment variables
python-dotenv==1.0.1

# Type checking
typing-extensions>=4.12.2

# Utilities
requests==2.32.3